Staging

+++ src/modules/workshop/UI/Panel/BottomBar/index.tsx
@@ -1,31 +1,37 @@
 'use client'
 import { IconClear, IconRedo, IconTrash, IconUndo } from '@/components/IconSvgs'
 import { useUndoRedoShortcut } from '@/hooks/useShortcuts'
+import IcCreate from '@/icons/workshop/ic-create.svg'
 import IcOpen from '@/icons/workshop/ic-open.svg'
-import { default as IcCreate, default as IcSave } from '@/icons/workshop/ic-save.svg'
-import { useProjectStore, useStoreGlobal } from '@/stores/blocks'
+import { default as IcSave } from '@/icons/workshop/ic-save.svg'
+import { useModalStore, useProjectStore, useStoreGlobal } from '@/stores/blocks'
 import s from './styles.module.scss'
 
 import useApiInfinite from '@/hooks/useApiInfinite'
-import { getListModularByWallet } from '@/services/api/generative'
-import { TBlockData, TListCurrent } from '@/types'
-import instance from '@/utils/storage/local-storage'
+import IcTwitter from '@/icons/workshop/ic-twitter.svg'
+import { getListModularByWallet, getProjectDetail, handleGetData, uploadFile } from '@/services/api/generative'
 import { useAppSelector } from '@/stores/hooks'
 import { accountSelector } from '@/stores/states/wallet/selector'
-import { handleConvertData } from '@/utils/convertTraits'
-import { useEffect } from 'react'
+import { TListCurrent } from '@/types'
+import { useEffect, useMemo, useRef } from 'react'
+
 
 type TDataFetch = {
   list: TListCurrent
 }
 
-import SavedProjectsModal from '@/modules/workshop/components/Modal/SavedProjectsModal'
+import { WORKSHOP_URL } from '@/constant/route-path'
+import SavedProjectsModal, { SAVED_PROJECTS_MODAL_ID } from '@/modules/workshop/components/Modal/SavedProjectsModal'
+import SetProjectNameModal, { SET_PROJECT_NAME_MODAL_ID } from '@/modules/workshop/components/Modal/SetProjectNameModal'
+import UnsaveWarningModal from '@/modules/workshop/components/Modal/UnsaveWarningModal'
+import { EDIT_MODE, captureCanvasImage } from '@/utils'
+import { convertBase64ToFile } from '@/utils/file'
+import { SHA256 } from 'crypto-js'
+import { useRouter } from 'next/navigation'
 import { useId, useState } from 'react'
 import { Toaster } from 'react-hot-toast'
-import jsonFile from './mock.json'
-import UnsaveWarningModal from '@/modules/workshop/components/Modal/UnsaveWarningModal'
 
-const MOCK_ADDRESS = 'bc1p4psqwcglffqz87kl0ynzx26dtxvu3ep75a02d09fshy90awnpewqvkt7er'
+// const MOCK_ADDRESS = 'bc1p4psqwcglffqz87kl0ynzx26dtxvu3ep75a02d09fshy90awnpewqvkt7er'
 
 export default function BottomBar() {
   const {
@@ -35,20 +41,35 @@ export default function BottomBar() {
     viewPreview,
     setViewPreview,
     deleteAlls,
-    blockCurrent,
     setDataCurrent,
     setBlockCurrent,
     blocksState,
+    selectedBricks,
     // deleteSeletBlocks,
+    deleteSelected,
+    setSelectedBricks,
+    setBricks,
+    blockCurrent,
+    setBlockCurrentUpdate,
   } = useStoreGlobal()
 
-  const { projectId, saveProject, createProject, projectName, renderFile } = useProjectStore()
+  const router = useRouter()
+
+  const { setLoading, projectId, saveProject, createProject, projectName, renderFile, loadProject } = useProjectStore()
+
+
+
+  const { openModal, modals } = useModalStore()
 
   const [showModal, setShowModal] = useState(false)
   const [showUnsaveModal, setShowUnsaveModal] = useState(false)
+  const [showSetProjectNameModal, setShowSetProjectNameModal] = useState(false)
+  const [clickView, setClickView] = useState(false)
+
+  const currentBlockStateRef = useRef(SHA256(JSON.stringify(blockCurrent)).toString() || '')
 
   const account = useAppSelector(accountSelector)
-
+  const isEditMode = mode === EDIT_MODE
   const id = useId()
 
   const {
@@ -58,7 +79,8 @@ export default function BottomBar() {
   } = useApiInfinite(
     getListModularByWallet,
     {
-      ownerAddress: MOCK_ADDRESS, //account?.address,
+      ownerAddress: account?.address,
+      // ownerAddress: 'bc1pafhpvjgj5x7era4cv55zdhpl57qvj0c60z084zsl7cwlmn3gq9tq3hqdmn',
       page: 1,
       limit: 20,
     },
@@ -76,20 +98,40 @@ export default function BottomBar() {
   const redoAction = () => {
     redo()
   }
-  const deleteAction = () => {
-    // deleteSeletBlocks()
-  }
+
+  const isAllowSave = useMemo(() => {
+    const hashBlockCurrent = SHA256(JSON.stringify(blockCurrent)).toString()
+
+    return hashBlockCurrent !== currentBlockStateRef.current
+  }, [blockCurrent])
+
   const saveAction = async () => {
-    if (blocksState.length < 2 || !blockCurrent || blockCurrent.length === 0) return
+
+    if (!isAllowSave) return
+
+    if (!projectName) {
+      openModal({
+        id: SET_PROJECT_NAME_MODAL_ID,
+        component: <SetProjectNameModal type='save' />,
+      })
+      return
+    }
+    setLoading(true)
+
+    const {dataURL: image} = captureCanvasImage({})
+    const file = convertBase64ToFile(image)
+    const resUrl = await uploadFile({ file })
 
     const payload: {
       jsonFile: any
       projectId?: string
       projectName?: string
       ownerAddress: string
+      thumbnail: string
     } = {
       jsonFile: blockCurrent,
-      ownerAddress: MOCK_ADDRESS, //account?.address,
+      ownerAddress: account?.address,
+      thumbnail: resUrl.url,
     }
 
     if (projectId) {
@@ -99,8 +141,19 @@ export default function BottomBar() {
     if (projectName) {
       payload.projectName = projectName
     }
+    await saveProject(payload)
 
-    saveProject(payload)
+    setLoading(false)
+  }
+
+  const saveAsAction = async () => {
+    if (!isAllowSave) return
+
+    openModal({
+      id: SET_PROJECT_NAME_MODAL_ID,
+      component: <SetProjectNameModal type='save-as' />,
+    })
+    return
   }
 
   const loadDataAction = (file) => {
@@ -115,27 +168,70 @@ export default function BottomBar() {
     deleteAlls()
   }
 
-  const handleGetData = async () => {
-    const data = (await getListModularByWallet({
-      ownerAddress: 'bc1pafhpvjgj5x7era4cv55zdhpl57qvj0c60z084zsl7cwlmn3gq9tq3hqdmn',
-      page: 1,
-      limit: 20,
-    })) as TDataFetch
-    const convertData =
-      Array.isArray(data.list) &&
-      data.list.map((item) => {
-        const traits = handleConvertData(item.attributes)
-        return { ...item, traits }
-      })
-    setDataCurrent(convertData)
-  }
+  // const handleGetData = async () => {
+  //   const data = (await getListModularByWallet({
+  //     ownerAddress: account?.address,
+  //     // ownerAddress: 'bc1pafhpvjgj5x7era4cv55zdhpl57qvj0c60z084zsl7cwlmn3gq9tq3hqdmn',
+  //     page: 1,
+  //     limit: 20,
+  //   })) as any
+  //   const listData = data.list as TListCurrent[]
+  //   return listData
+  // }
 
-  const handleClickCreateNewProject = () => {
-    // if (blocksState.length > 2) {
-    //   setShowUnsaveModal(true)
-    // }
+  const handleClickCreateNewProject = async () => {
+    if (isAllowSave) {
+      setShowUnsaveModal(true)
+      return
+    }
     createProject()
     deleteAlls()
+    const data = await handleGetData(account?.address) //reset new project
+    setDataCurrent(data)
+  }
+
+  const viewAction = async () => {
+    if (!projectId && !isAllowSave) return
+
+    if (!projectId && isAllowSave) {
+      openModal({
+        id: SET_PROJECT_NAME_MODAL_ID,
+        component: <SetProjectNameModal type='save-view' />,
+      })
+      return
+    }
+
+    if (isAllowSave) {
+      await saveAction()
+      setLoading(true)
+      setTimeout(() => {
+        window.open(`${WORKSHOP_URL}/${projectId}`, '_blank')
+      }, 3000)
+      return
+    }
+    window.open(`${WORKSHOP_URL}/${projectId}`, '_blank')
+
+  }
+
+  const loadInitialProject = async () => {
+    try {
+      setLoading(true)
+      const res = await getProjectDetail({ id: projectId })
+      // const data = await handleGetData(account.address) // reset data when open new data
+
+      if (!!res.metaData) {
+        // setDataCurrent(data)
+        loadProject({
+          projectId: projectId,
+          projectName: projectName,
+          renderFile: res.metaData,
+        })
+      }
+    } catch (error) {
+      //
+    } finally {
+      setLoading(false)
+    }
   }
 
   useUndoRedoShortcut(undo, redo)
@@ -147,57 +243,83 @@ export default function BottomBar() {
   }, [renderFile])
 
   useEffect(() => {
-    handleGetData()
-  }, [])
+    // detect click browser back button or closing tab
+    if (!isAllowSave) return
+    window.addEventListener('beforeunload', (e) => {
+      e.preventDefault()
+    })
 
-  // First load, trigger create new project
-  useEffect(() => {
-    if (!projectId) {
-      createProject()
+    return () => {
+      window.removeEventListener('beforeunload', (e) => {
+        e.preventDefault()
+      })
     }
-  }, [projectId, createProject])
+  }, [isAllowSave])
+
+  useEffect(() => {
+    if (projectId && projectName) {
+      loadInitialProject()
+    }
+  }, [projectId, projectName])
+
+
+
+  const handleDeleteSelected = () => {
+    deleteSelected(selectedBricks)
+    setSelectedBricks({})
+  }
 
   return (
     <>
       <Toaster />
       <div className={s.wrapper}>
         <div className={s.bottomBar}>
-          <button className={s.bottomBar_btn} onClick={undoAction}>
-            <IconUndo /> Undo
+          <button className={`${s.bottomBar_btn} ${s.noFill}`} onClick={undoAction}>
+            <IconUndo />
           </button>
-          <button className={s.bottomBar_btn} onClick={redoAction}>
-            <IconRedo /> Redo
+          <button className={`${s.bottomBar_btn} ${s.noFill}`} onClick={redoAction}>
+            <IconRedo />
           </button>
           <button className={s.bottomBar_btn} onClick={() => handleDeleteAll()}>
             <IconClear />
-            Clear
           </button>
-          <button className={s.bottomBar_btn} onClick={deleteAction}>
-            <IconTrash /> Delete
-          </button>
+          {isEditMode && (
+            <button
+              className={`${s.bottomBar_btn} ${selectedBricks.length === 0 && s.disable}`}
+              disabled={selectedBricks.length === 0 && true}
+              onClick={handleDeleteSelected}
+            >
+              <IconTrash />
+            </button>
+          )}
         </div>
 
         <div className={s.bottomBar}>
-          <button className={s.bottomBar_btn} onClick={saveAction}>
-            <IcSave /> Save Project
+          <button className={`${s.bottomBar_btn} ${s.icon_X}`} onClick={viewAction}>
+            <IcTwitter /> Share
           </button>
-          <button className={s.bottomBar_btn} onClick={saveAction}>
+          <button className={s.bottomBar_btn} onClick={saveAction} disabled={!isAllowSave}>
+            <IcSave /> Save
+          </button>
+          {/* <button className={s.bottomBar_btn} onClick={saveAsAction} disabled={!isAllowSave}>
             <IcSave /> Save As
-          </button>
+          </button> */}
           <button className={s.bottomBar_btn} onClick={handleClickCreateNewProject}>
-            <IcCreate /> Create New
+            <IcCreate /> New
           </button>
           <button
             className={s.bottomBar_btn}
             onClick={() => {
-              setShowModal(true)
+              openModal({
+                id: SAVED_PROJECTS_MODAL_ID,
+                component: <SavedProjectsModal />,
+              })
             }}
           >
-            <IcOpen /> Open Project
+            <IcOpen /> Open
           </button>
         </div>
       </div>
-      <SavedProjectsModal show={showModal} setIsOpen={setShowModal} />
       <UnsaveWarningModal show={showUnsaveModal} setIsOpen={setShowUnsaveModal} />
     </>
   )
